requirements.txt
django == 4.1
djangorestframework
psycopg2-binary
gunicorn
pillow
requests
django-debug-toolbar
PyJWT